+++ platform/application/platform/extensionsmanager.php
@@ -171,32 +171,34 @@ class ExtensionsManager
 
         // Register this extension routes.
         //
-        if ( array_key_exists('routes', $extension) )
+        if ( $routes = array_get($extension, 'routes') )
         {
             // Check if we've been given a closure.
-
-            if ( ! $extension['routes'] instanceof Closure )
+            //
+            if ( ! $routes instanceof Closure )
             {
                 throw new Exception('"routes" must be a function / closure in [' . $slug . ']');
             }
 
-            $extension['routes']();
+            // Register it.
+            //
+            $routes();
         }
 
         // Register this extension listeners.
         //
-        if ( array_key_exists('listeners', $extension) )
+        if ( $listeners = array_get($extension, 'listeners') )
         {
             // Check if we've been given a closure.
             //
-            if ( ! $extension['listeners'] instanceof Closure )
+            if ( ! $listeners instanceof Closure )
             {
                 throw new Exception('"listeners" must be a function / closure in [' . $slug . ']');
             }
 
             // Register it.
             //
-            $extension['listeners']();
+            $listeners();
         }
 
         // The extension has been started.
@@ -240,7 +242,7 @@ class ExtensionsManager
             //
             $slug = $extension['info']['slug'];
 
-             // Check if this extension is installed.
+            // Check if this extension is installed.
             //
             if ( $info = array_get( $this->installed, $slug ) )
             {
@@ -268,7 +270,7 @@ class ExtensionsManager
                     $extension['info']['can_enable'] = $this->can_enable( $slug );
                 }
 
-                // Sort it.
+                // Sort the info array.
                 //
                 ksort( $extension['info'] );
 
@@ -283,7 +285,7 @@ class ExtensionsManager
                 //
                 $extension['info']['can_install'] = $this->can_install( $slug );
 
-                // Sort it.
+                // Sort the info array.
                 //
                 ksort( $extension['info'] );
 
@@ -425,37 +427,6 @@ class ExtensionsManager
         return $this->enabled = $this->installed(function($query){
             return $query->where('enabled', '=', '1');
         });
-/*
-        // Initiate an empty array.
-        //
-        $extensions = array();
-
-        // Spin through the enabled extensions.
-        //
-        foreach ( Extension::where_enabled(1)->get() as $extension )
-        {
-            // Extension slug.
-            //
-            $slug = $extension->slug;
-
-            // Extension information.
-            //
-            $extensions[ $slug ] = array_replace_recursive( $this->extensions[ $slug ], array(
-                'slug'    => $slug,
-                'version' => $extension->version,
-                //'is_core' => (bool) $extension->is_core,
-                'enabled' => (bool) $extension->enabled
-            ) );
-        }
-
-        // Sort the extensions.
-        //
-        ksort( $extensions );
-
-        // Store and return the extensions.
-        //
-        return $this->enabled = $extensions;
-*/
     }
 
 
@@ -483,37 +454,6 @@ class ExtensionsManager
         return $this->disabled = $this->installed(function($query){
             return $query->where('enabled', '=', '1');
         });
-/*
-        // Initiate an empty array.
-        //
-        $extensions = array();
-
-        // Spin through the disabled extensions.
-        //
-        foreach ( Extension::where_enabled(0)->get() as $extension )
-        {
-            // Extension slug.
-            //
-            $slug = $extension->slug;
-
-            // Extension information.
-            //
-            $extensions[ $slug ] = array_replace_recursive( $this->extensions[ $slug ], array(
-                'slug'    => $slug,
-                'version' => $extension->version,
-                'is_core' => (bool) $extension->is_core,
-                'enabled' => (bool) $extension->enabled
-            ) );
-        }
-
-        // Sort the extensions.
-        //
-        ksort( $extensions );
-
-        // Store and return the extensions.
-        //
-        return $this->disabled = $extensions;
-*/
     }
 
 
@@ -1531,249 +1471,7 @@ class ExtensionsManager
         //
         $this->start_extensions();
     }
-
-
-
-
-    /**
-     * Sorts extension dependencies
-     *
-     * @param   array  $extensions
-     * @return  array  $extensions
-     */
-    public function sort_dependencies(&$extensions = array())
-    {
-        // Array of extensions dependencies, where
-        // the key is the slug of the extension
-        // and the value is an array of extension slugs
-        // on which that extension depends.
-        $extensions_dependencies = array();
-
-        foreach ($extensions as $extension)
-        {
-            $slug = $extension['info']['slug'];
-            // try
-            // {
-            //  $extension = $this->get_extensionphp($slug);
-            // }
-            // catch (Exception $e)
-            // {
-            //  continue;
-            // }
-
-            if ($dependencies = array_get($extension, 'dependencies') and is_array($dependencies))
-            {
-                $extensions_dependencies[$slug] = $dependencies;
-            }
-            else
-            {
-                $extensions_dependencies[$slug] = array();
-            }
-        }
-
-        return Dependency::sort($extensions_dependencies);
-    }
 }
 
 /* End of file extensionsmanager.php */
 /* Location: ./platform/application/platform/extensionsmanager.php */
-
-
-
-/**
- * @todo Maybe put this in it's own file...
- */
-
-class Dependency
-{
-    public static function sort($extensions = array())
-    {
-        // The class below requires that we have
-        // at least 1 dependency for each module.
-        foreach ($extensions as $extension => &$dependencies)
-        {
-            if (empty($dependencies))
-            {
-                $dependencies[] = 'core';
-            }
-        }
-
-        $t = new TopologicalSort($extensions, true);
-        $sorted = $t->tsort();
-
-        if ( ! $sorted)
-        {
-            throw new Exception('Error in sorting dependencies');
-        }
-
-        // Search for core (the most basic placehodler
-        // dependency we provided)
-        if (in_array('core', $sorted))
-        {
-            // Try keep keys sorted nicely
-            if (($key = array_search('core', $sorted)) === 0)
-            {
-                array_shift($sorted);
-            }
-            else
-            {
-                unset($sorted[array_search('core', $sorted)]);
-            }
-        }
-
-        return $sorted;
-    }
-}
-
-
-/**
- * @todo refactor and implement the below class proprly.
- */
-
-
-
-
-
-/**
-* Sorts a series of dependency pairs in linear order
-*
-* usage:
-* $t = new TopologicalSort($dependency_pairs);
-* $load_order = $t->tsort();
-*
-* where dependency_pairs is in the form:
-* $name => (depends on) $value
-*
-*/
-class TopologicalSort
-{
-    public $nodes = array();
-
-    /**
-    * Dependency pairs are a list of arrays in the form
-    * $name => $val where $key must come before $val in load order.
-    *
-    */
-    public function __construct($dependencies=array(), $parse=false)
-    {
-        if ($parse) $dependencies = $this->parseDependencyList($dependencies);
-        // turn pairs into double-linked node tree
-
-        foreach($dependencies as $key => $dpair) {
-            list($module, $dependency) = each($dpair);
-
-            if (! isset($this->nodes[$module]))
-                $this->nodes[$module] = new TSNode($module);
-
-            if (! isset($this->nodes[$dependency]))
-                $this->nodes[$dependency] = new TSNode($dependency);
-
-            if (! in_array($dependency,$this->nodes[$module]->children))
-                $this->nodes[$module]->children[] = $dependency;
-
-            if (! in_array($module,$this->nodes[$dependency]->parents))
-                $this->nodes[$dependency]->parents[] = $module;
-        }
-    }
-
-    /**
-    * Perform Topological Sort
-    *
-    * @param array $nodes optional array of node objects may be passed.
-    * Default is  $this->nodes created in constructor.
-    * @return sorted array
-    */
-    public function tsort($nodes=array())
-    {
-        // use this->nodes if it is populated and no param passed
-        if (! @count($nodes) && count($this->nodes))
-        $nodes = $this->nodes;
-
-        // get nodes without parents
-        $root_nodes = array_values($this->getRootNodes($nodes));
-
-        // begin algorithm
-        $sorted = array();
-        while(count($nodes)>0) {
-            // check for circular reference
-            if (count($root_nodes) == 0) return false;
-
-            // remove this node from root_nodes
-            // and add it to the output
-            $n = array_pop($root_nodes);
-            $sorted[] = $n->name;
-
-            // for each of its  children
-            // queue the new node finally remove the original
-            for($i=(count($n->children)-1); $i >= 0; $i--) {
-                $childnode = $n->children[$i];
-                // remove the link from this node to its
-                // children ($nodes[$n->name]->children[$i]) AND
-                // remove the link from each child to this
-                // parent ($nodes[$childnode]->parents[?]) THEN
-                // remove this child from this node
-                unset($nodes[$n->name]->children[$i]);
-                $parent_position = array_search($n->name,$nodes[$childnode]->parents);
-                unset($nodes[$childnode]->parents[$parent_position]);
-                // check if this child has other parents
-                // if not, add it to the root nodes list
-                if (!count($nodes[$childnode]->parents))array_push($root_nodes,$nodes[$childnode]);
-            }
-
-            // nodes.Remove(n);
-            unset($nodes[$n->name]);
-        }
-        return $sorted;
-    }
-
-    /**
-    * Returns a list of node objects that do not have parents
-    *
-    * @param array $nodes array of node objects
-    * @return array of node objects
-    */
-    public function getRootNodes($nodes)
-    {
-    $output = array();
-    foreach($nodes as $name => $node)
-     if (!count($node->parents)) $output[$name] = $node;
-    return $output;
-    }
-
-    /**
-    * Parses an array of dependencies into an array of dependency pairs
-    *
-    * The array of dependencies would be in the form:
-    * $dependency_list = array(
-    *  "name" => array("dependency1","dependency2","dependency3"),
-    *  "name2" => array("dependencyA","dependencyB","dependencyC"),
-    *  ...etc
-    * );
-    *
-    * @param array $dlist Array of dependency pairs for use as parameter in tsort method
-    * @return array
-    */
-    public function parseDependencyList($dlist=array())
-    {
-    $output = array();
-        foreach($dlist as $name => $dependencies)
-            foreach($dependencies as $d)
-                array_push($output, array($d => $name));
-        return $output;
-    }
-}
-
-/**
-* Node class for Topological Sort Class
-*
-*/
-class TSNode
-{
-    public $name;
-    public $children = array();
-    public $parents = array();
-
-    public function __construct($name="") {
-        $this->name = $name;
-    }
-}